fix: syscall dispatcher return type

+++ Kernel/interrupts/syscalls/scDispatcher.c
@@ -4,7 +4,7 @@
 #include <scDispatcher.h>
 #include <syscalls.h>
 
-uint8_t syscall_dispatcher(uint64_t arg0, uint64_t arg1, uint64_t arg2,
+int syscall_dispatcher(uint64_t arg0, uint64_t arg1, uint64_t arg2,
                            uint64_t arg3, uint64_t arg4, uint64_t arg5,
                            uint64_t id) {
     switch (id) {
